requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1


google-generativeai==0.7.2
httpx==0.27.2

qdrant-client==1.12.1
numpy==1.26.4
